UBUNTU: SAUCE: overlayfs: when copying up and reading directories ensure mounter had permissions V2
When copying up files and reading directory contents ensure the mounter has
permissions for the operation over the consituent parts (lower and upper).
Where we are in a namespace this ensures that the mounter (root in that
namespace) has permissions over the files and directories, preventing
exposure of protected files and directory contents.

CVE-2015-1328
Acked-by: Tyler Hicks <tyhicks@canonical.com>
Acked-by: Serge E. Hallyn <serge.hallyn@ubuntu.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Luis Henriques <luis.henriques@canonical.com>
(backported from commit 2fc940cd3ddb43415422ed771173b38685c2dbf9 in vivid/Ubuntu-3.19.0-21.21)
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>

Conflicts:
	fs/overlayfs/copy_up.c
	fs/overlayfs/readdir.c
	fs/overlayfs/super.c