BCM270X_DT: Use fixed-factor-clock for uart1
The BCM2835 mini-UART is almost 8250-compatible, but there is
a factor-of-two difference in the clock divider for the baud rate.
The standard 8250 driver can be used unmodified provided we lie to it
about the clock frequency, pretending it is double the actual value.
The mini-UART block shares the core clock, and the current firmware
uses a DT-parameter to update the clock-frequency property. Although it
works, this isn't ideal because it exposes the firmware to a driver
problem.

This patch uses the fixed-factor-clock module to create a clock that is
always twice the frequency of the core clock. (N.B. These "fixed" clocks
just describe an existing clock, rather than changing any hardware
settings.)

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>