UBUNTU: SAUCE: KVM: arm/arm64: add emulation model specific destroy function
Currently we destroy the VGIC emulation in one function that cares for
all emulated models. To be on par with init_model (which is model
specific), lets introduce a per-emulation-model destroy method, too.
Use it for a tiny GICv3 specific code already, later it will be handy
for the ITS emulation.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Reviewed-by: Eric Auger <eric.auger@linaro.org>
[ From: http://www.spinics.net/lists/kvm/msg118908.html ]
Signed-off-by: dann frazier <dann.frazier@canonical.com>

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>