UBUNTU: SAUCE: KVM: arm64: Introduce new MMIO region for the ITS base address
The ARM GICv3 ITS controller requires a separate register frame to
cover ITS specific registers. Add a new VGIC address type and store
the address in a field in the vgic_dist structure.
Provide a function to check whether userland has provided the address,
so ITS functionality can be guarded by that check.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
[ From: http://www.spinics.net/lists/kvm/msg118907.html ]
Signed-off-by: dann frazier <dann.frazier@canonical.com>

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>