UBUNTU: SAUCE: Add seccomp selftests.
BugLink: http://bugs.launchpad.net/bugs/1509489

Taken from the history of https://github.com/redpig/seccomp. This
has been deprecated in favor of the tests in mainline (which are
based on this repository), but the mainline tests include tests
for the seccomp syscall and tsync, neither of which are included
in this backport.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>
Acked-by: Kamal Mostafa <kamal@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>